code refactoring part 2
Simplified code in 4 of 6 .JS files. Most changes were in the ui.js. Instead of creating the element and setting it's class and ID and then appending it for every ui element, this process is done by a function called "createUiElement".

In the worldGrid.js file unnecessary code was removed and a world generation test was removed.

In the moveByVector.js file the object movement function now calls the entity movement function for actually moving and just cleans the previous representation in the worldGridData.

In the core.js file, the "stats_data" was removed, now all info of stats shall be taken from the "ui_data". Note that the "clothing" and "equiped" properties are not affected but the "health", "hunger", "thirst" and "sanity" need to be calculated by getting the length of the strings. Also in the core.js file, the "updateWorldGrid" call was updated to reflect previously mentioned changes and FPS tracking was removed.