added behaviors to objects + code refactoring
Behaviors are functions that take the gameState const and the object they should update and are run every frame.
To add these behaviors to an object it's just needed to set the "behavior" property of the object to the name of the behavior function. Two behaviors are currently implemented in the game. 1º Is the player movement behavior which listens to the keyboard to generate the movement vector and when can move(movement cool down has passed) moves the player object by the generated vector. 2º Is a test for zombie behavior, the zombie just moves to the left of the screen.

** The camera entity was completely removed from the game and now the player is passed as a camera, making the code simpler.